py-bot/requirements.txt
discord.py
python-dotenv
websockets